+++ lib/Secucard.Connect/Product/Payment/Model/SecupayCreditcard.cs
@@ -5,14 +5,9 @@
     [DataContract]
     public class SecupayCreditcard : Transaction
     {
-        [DataMember(Name = "container")]
-        public Container Container { get; set; }
-
         public override string ToString()
         {
-            return "SecupayCreditcard{" +
-                   "container=" + Container +
-                   "} " + base.ToString();
+            return "SecupayCreditcard{ " + base.ToString() + " }";
         }
     }
 }